+++ test-harness/index.ts
@@ -1,20 +1,24 @@
-import { parseSpecFile } from './helpers';
+import { ChildProcess, spawn, spawnSync } from 'child_process';
 import * as fs from 'fs';
+import { validate } from 'gavel';
+import * as globFs from 'glob-fs';
+import { parseResponse } from 'http-string-parser';
 import * as os from 'os';
 import * as path from 'path';
-import * as tmp from 'tmp';
-import { ChildProcess, spawnSync, spawn } from 'child_process';
 import * as split2 from 'split2';
-const validate = require('gavel').validate;
-import { parseResponse } from 'http-string-parser';
+import * as tmp from 'tmp';
+import { parseSpecFile } from './helpers';
 
-jest.setTimeout(60000);
+const glob = globFs({ gitignore: true });
+jest.setTimeout(5000);
 
 describe('harness', () => {
-  const files = fs.readdirSync(path.join(__dirname, './specs/'));
+  const files = process.env.TESTS
+    ? String(process.env.TESTS).split(',')
+    : glob.readdirSync('**/*.txt', { cwd: path.join(__dirname, './specs') });
 
-  files.forEach(value => {
-    const data = fs.readFileSync(path.join(__dirname, './specs/', value), { encoding: 'utf8' });
+  files.forEach(file => {
+    const data = fs.readFileSync(path.join(__dirname, './specs/', file), { encoding: 'utf8' });
     const parsed = parseSpecFile(data);
 
     let prismMockProcessHandle: ChildProcess;
@@ -36,7 +40,7 @@ describe('harness', () => {
 
     afterAll(() => tmpFileHandle.removeCallback(undefined, undefined, undefined, undefined));
 
-    test(`${value}${os.EOL}${parsed.test}`, done => {
+    test(`${file}${os.EOL}${parsed.test}`, done => {
       const [command, ...args] = parsed.command.split(' ').map(t => t.trim());
       const serverArgs = [...parsed.server.split(' ').map(t => t.trim()), tmpFileHandle.name];
 
@@ -50,9 +54,7 @@ describe('harness', () => {
             windowsVerbatimArguments: false,
           });
           const output: any = parseResponse(clientCommandHandle.stdout.trim());
-          const expected: any = parseResponse(
-            parsed.expect.trim() || parsed.expectLoose.trim()
-          );
+          const expected: any = parseResponse(parsed.expect.trim() || parsed.expectLoose.trim());
 
           try {
             const isValid = validate(expected, output).valid;